backend/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wheel==0.45.1
